MSVC compilation fixes, replacing wingetopt with Kim Grasman's BSD-licensed version: https://github.com/kimgr/getopt_port

+++ src/applications/staticProps/Field.cpp
@@ -242,9 +242,9 @@ namespace OpenMD {
       }
       
       // find xyz-indices of cell that stuntDouble is in.
-      ibin = nBins_.x() * scaled.x();
-      jbin = nBins_.y() * scaled.y();
-      kbin = nBins_.z() * scaled.z();
+      ibin = (int) (nBins_.x() * scaled.x());
+      jbin = (int) (nBins_.y() * scaled.y());
+      kbin = (int) (nBins_.z() * scaled.z());
       
       for (int i = -di; i <= di; i++) {
 	igrid = ibin + i;
